Deleted SensorArray files and added the sensor code to the main file. Subject to revision

+++ Wifi_Sensor_WIP/WifiConnection.h
@@ -62,7 +62,7 @@ public:
    * A five second interval between connection information displays
    * ------ MAY NOT BE NECESSARY ------
    */
-  const int intervalInfo = 5000;
+  const long intervalInfo = 5000;
 
   /*
    * A flag to display connection information
